A decoder for the Vigenere Cipher
A decoder for the Vigenere Cipher (English Language). First the IC is calculated for a number of values, estimating the key length. Then we use the frequencies of letters in the English Language we retrieve the suspected key and lastly, we return the (decoded) message. The ciphertext is given as an argument